ScrappingMaroc/requirements.txt
beautifulsoup4
selenium
webdriver-manager
pymongo
python-dotenv
certifi